Split datetime-related utility functions into a separate module
