+++ requests_cache/_utils.py
@@ -1,7 +1,7 @@
-"""Miscellaneous minor utility functions that don't really belong anywhere else"""
+"""Minor internal utility functions that don't really belong anywhere else"""
 from inspect import signature
 from logging import getLogger
-from typing import Any, Callable, Dict, Iterable, Iterator, List
+from typing import Any, Callable, Dict, Iterable, Iterator, List, Optional
 
 logger = getLogger('requests_cache')
 
@@ -16,6 +16,18 @@ def chunkify(iterable: Iterable, max_size: int) -> Iterator[List]:
 def coalesce(*values: Any, default=None) -> Any:
     """Get the first non-``None`` value in a list of values"""
     return next((v for v in values if v is not None), default)
+
+
+def decode(value, encoding='utf-8') -> str:
+    """Decode a value from bytes, if hasn't already been.
+    Note: ``PreparedRequest.body`` is always encoded in utf-8.
+    """
+    return value.decode(encoding) if isinstance(value, bytes) else value
+
+
+def encode(value, encoding='utf-8') -> bytes:
+    """Encode a value to bytes, if it hasn't already been"""
+    return value if isinstance(value, bytes) else str(value).encode(encoding)
 
 
 def get_placeholder_class(original_exception: Exception = None):
@@ -46,3 +58,11 @@ def get_valid_kwargs(func: Callable, kwargs: Dict, extras: Iterable[str] = None)
     params = list(signature(func).parameters)
     params.extend(extras or [])
     return {k: v for k, v in kwargs.items() if k in params and v is not None}
+
+
+def try_int(value: Any) -> Optional[int]:
+    """Convert a value to an int, if possible, otherwise ``None``"""
+    try:
+        return int(value)
+    except (TypeError, ValueError):
+        return None
